chromium/third_party/blink/web_tests/external/wpt/scroll-animations/css/animation-timeline-in-keyframe-expected.txt

This is a testharness.js-based test.
FAIL The animation-timeline property may not be used in keyframes assert_equals: expected (string) "auto" but got (undefined) undefined
Harness: the test ran to completion.
